requirements.txt
a2s==1.0.1
Pillow==9.3.0
Pillow==7.0.0
Pillow==9.5.0
pyTelegramBotAPI==4.7.1
pyTelegramBotAPI==4.11.0
pytesseract==0.3.10
python_a2s==1.3.0
Requests==2.30.0
schedule==1.2.0
